refactor(formDesigner): moved setActionFlag from from provider
