+++ shesha-reactjs/src/providers/form/index.tsx
@@ -1,5 +1,5 @@
 import { FormInstance } from 'antd';
-import React, { FC, MutableRefObject, PropsWithChildren, ReactNode, useContext, useEffect, useMemo } from 'react';
+import React, { FC, MutableRefObject, PropsWithChildren, useContext, useEffect, useMemo } from 'react';
 import { useDeepCompareEffect } from 'react-use';
 import { useDebouncedCallback } from 'use-debounce';
 import useThunkReducer from '../../hooks/thunkReducer';
@@ -16,13 +16,11 @@ import { useGlobalState } from '../globalState';
 import { getFlagSetters } from '../utils/flagsSetters';
 import {
   registerComponentActionsAction,
-  setActionFlagAction,
   setEnabledComponentsAction,
   setFormControlsDataAction,
   setFormDataAction,
   setFormModeAction,
   setSettingsAction,
-  setToolbarRightButtonAction,
   setValidationErrorsAction,
   setVisibleComponentsAction,
 } from './actions';
@@ -39,7 +37,7 @@ import {
   ISetVisibleComponentsPayload,
 } from './contexts';
 import { useFormDesignerComponents } from './hooks';
-import { FormMode, FormRawMarkup, IFormActions, IFormDesignerActionFlag, IFormSections, IFormSettings } from './models';
+import { FormMode, FormRawMarkup, IFormActions, IFormSections, IFormSettings } from './models';
 import formReducer from './reducer';
 import { convertActions, convertSectionsToList, getEnabledComponentIds, getVisibleComponentIds } from './utils';
 import { useDataContextManager } from 'providers/dataContextManager';
@@ -215,7 +213,7 @@ const FormProvider: FC<PropsWithChildren<IFormProviderProps>> = ({
   const isComponentDisabled = (model: Pick<IConfigurableFormComponent, 'id' | 'isDynamic' | 'disabled'>): boolean => {
     const disabledByCondition = false;
     //ToDo AS: need to change the way to update visible end enabled comopnents
-      //model.isDynamic !== true && state.enabledComponentIds && !state.enabledComponentIds.includes(model.id);
+    //model.isDynamic !== true && state.enabledComponentIds && !state.enabledComponentIds.includes(model.id);
 
     return state.formMode !== 'designer' && (model.disabled || disabledByCondition);
   };
@@ -224,7 +222,7 @@ const FormProvider: FC<PropsWithChildren<IFormProviderProps>> = ({
     const hiddenByCondition = false;
 
     //ToDo AS: need to change the way to update visible end enabled comopnents
-      //model.isDynamic !== true && state.visibleComponentIds && !state.visibleComponentIds.includes(model.id);
+    //model.isDynamic !== true && state.visibleComponentIds && !state.visibleComponentIds.includes(model.id);
 
     return state.formMode !== 'designer' && (model.hidden || hiddenByCondition);
   };
@@ -414,14 +412,6 @@ const FormProvider: FC<PropsWithChildren<IFormProviderProps>> = ({
     return visibleChildIndex !== -1;
   };
 
-  const setActionFlag = (payload: IFormDesignerActionFlag) => {
-    dispatch(setActionFlagAction(payload));
-  };
-
-  const setToolbarRightButton = (payload: ReactNode) => {
-    dispatch(setToolbarRightButtonAction(payload));
-  };
-
   const configurableFormActions: IFormActionsContext = {
     ...getFlagSetters(dispatch),
     getComponentModel,
@@ -440,8 +430,6 @@ const FormProvider: FC<PropsWithChildren<IFormProviderProps>> = ({
     getToolboxComponent,
     setFormDataAndInstance,
     hasVisibleChilds,
-    setActionFlag,
-    setToolbarRightButton,
   };
   if (formRef) formRef.current = { ...configurableFormActions, ...state, allComponents, componentRelations };
