+++ shesha-reactjs/src/providers/formDesigner/models.ts
@@ -1,4 +1,6 @@
-import { IPropertyMetadata } from "../../interfaces/metadata";
+import { IPropertyMetadata } from '../../interfaces/metadata';
+
+export type IFormDesignerActionFlag = 'done' | 'redo' | 'publish' | 'settings' | 'undo' | 'version';
 
 /** Named Data Source */
 export interface IDataSource {
